refactor(Ontologies): adapted Ontologies interface and classes to import schema
