+++ src/model/interfaces/index.ts
@@ -3,4 +3,5 @@ export {Entity} from "./Entity";
 export {Property} from "./Property";
 export {Individual} from "./Individual";
 export {Ontology} from "./Ontology";
+export {Ontologies} from "./Ontologies"
 export {Thing} from "./Thing";
