+++ src/model/ols3-model/index.ts
@@ -2,5 +2,6 @@ export {OLS3Class} from "./OLS3Class";
 export {OLS3Property} from "./OLS3Property";
 export {OLS3Individual} from "./OLS3Individual";
 export {OLS3Ontology} from "./OLS3Ontology";
+export {OLS3Ontologies} from "./OLS3Ontologies";
 export {OLS3Entity} from "./OLS3Entity";
 export {OLS3Thing} from "./OLS3Thing";
