+++ src/model/ols4-model/index.ts
@@ -2,5 +2,6 @@ export {OLS4Class} from "./OLS4Class";
 export {OLS4Property} from "./OLS4Property";
 export {OLS4Individual} from "./OLS4Individual";
 export {OLS4Ontology} from "./OLS4Ontology";
+export {OLS4Ontologies} from "./OLS4Ontologies";
 export {OLS4Entity} from "./OLS4Entity";
 export {OLS4Thing} from "./OLS4Thing";
